% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accuracy.scm.R
\name{accuracy.scm}
\alias{accuracy.scm}
\title{Add accuracy metrics to an scm object}
\usage{
accuracy.scm(scm)
}
\arguments{
\item{scm}{The \code{scm} object to enhance.}
}
\value{
An enhanced \code{scm} object, with additional items:
totals for rows, columns, and overall for both the \code{P} matrix
and the statistics' associated uncertainties
(\code{P_row_totals}, \code{P_col_totals}, \code{P_total},
\code{U_row_totals}, \code{U_col_totals}, \code{U_total}),
users' and producers' accuracies (\code{P_user_accuracy}, \code{U_user_accuracy},
\code{P_producer_accuracy}, \code{U_producer_accuracy}), overall accuracies
(\code{P_overall_accuracy}, \code{U_overall_accuracy}) and kappa statistics
(\code{P_kappa}, \code{U_kappa}).
}
\description{
An accessory function that enhances an \code{scm} object to include
accuracy statistics. This function does not modify the \code{P} and \code{U}
matrices, making them compatible with \code{scm} objects without enhancement.
}
