% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_dll_info.R
\name{boost_eff}
\alias{boost_eff}
\title{Extract information from the boosting procedure}
\usage{
boost_eff(data_boost)
}
\arguments{
\item{data_boost}{an object obtained by ...}
}
\value{
The function
}
\description{
This function allows to obtain a summary of the effects chosen by the boosting
}
\examples{

}
