% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gb_order.R
\name{boost_order}
\alias{boost_order}
\title{Perform the gradient boosting}
\usage{
boost_order(
  y,
  X,
  effects,
  constraints,
  nstep = 1,
  verbose = FALSE,
  rate = 1,
  ncores = 1
)
}
\arguments{
\item{y}{outcome (residual)}

\item{X}{covariates}

\item{effects}{list of effects included in the boosting procedure}

\item{constraints}{list of constraints on the effects}

\item{nstep}{numbers of step of the boosting procedure}

\item{verbose}{to print the iteration summary of the procedure}

\item{rate}{learning rate}

\item{ncores}{number of cores for parallel computation}
}
\value{
A list of summaries
}
\description{
This function allows to obtain an ordering of importance of the effects on the unconstrained elements of the modified cholesky decomposition
}
\examples{

}
