% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Comparator.R
\name{Comparator}
\alias{Comparator}
\title{Comparator function for subpixel confusion matrices}
\usage{
Comparator(s_k, r_l, A = min, D = PROD_D, scale = FALSE)
}
\arguments{
\item{s_k}{A matrix or another object coercible to one of predicted values.
Each row must sum up to 1.}

\item{r_l}{A matrix or another object coercible to one of true observed values against which
the predicted values will be validated.}

\item{A}{A function for calculating the diagonals of the confusion matrix.
This can be any unary reduction function, such as \code{min}, taking two
numbers and returning one.}

\item{D}{A function for calculating the off-diagonals of the confusion
matrix. The function must take four arguments: the overestimation matrix,
the underestimation matrix, and indices for the row and column of the result.
With a custom function, the uncertainty is not calculated
(a zero matrix is returned).}

\item{scale}{Boolean, whether the result should be scaled (reported in percentages).}
}
\value{
A confusion matrix.
}
\description{
The low-level function used to calculate an \code{SCM}.
Used internally by \code{SCM}.
Should normally not be called directly unless by experienced users.
}
