% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula_mcd_gb_selection.R
\name{formula_mcd}
\alias{formula_mcd}
\title{Building formula for the covariance model}
\usage{
formula_mcd(data_boost, res, stop_elem = 20)
}
\arguments{
\item{data_boost}{The gradient boosting object}

\item{res}{The summary of the gradient boosting}

\item{stop_elem}{The number of elements considered for modelling purposes}
}
\value{
The formula (a list) for the covariance model
}
\description{
This function return the formula for the covariance model, namely the formula for the elements of the modified cholesky decomposition
}
\examples{

}
