% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_formula_mcd.R
\name{formula_mcd_int}
\alias{formula_mcd_int}
\title{covariance model initialization formula}
\usage{
formula_mcd_int(d)
}
\arguments{
\item{d}{The dimension of the outcome}
}
\value{
The initializing formula (a list) for the covariance model
}
\description{
This function return the covariance model initialization formula
}
\examples{

}
