% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam_scm.R, R/gam_scm2.R
\name{gam_scm}
\alias{gam_scm}
\alias{gam_scm2}
\title{Fit a multivariate GAM model with covariance modelling}
\usage{
gam_scm(
  formula,
  family = mvn_scm(d = 2, nb = 1, param = NULL),
  optimizer = NULL,
  data = list(),
  aGam = list()
)

gam_scm2(
  formula,
  family = mvn_scm(d = 2, nb = 1, param = NULL),
  optimizer = NULL,
  data = list(),
  auxGam = list()
)
}
\arguments{
\item{formula}{formula provided by the user}

\item{family}{set the family to mvn_scm (whose parameters are d - dimension of the outcome; nb - number of observations' blocks; param - the type of parametrisation: cuurently only mcd (default) and logm are available)}

\item{data}{same arguments as in \link[mgcv:gam]{mgcv::gam}}

\item{auxGam}{list of further arguments to be passed to \link[mgcv:gam]{mgcv::gam}}
}
\value{
An object of class "scm"

An object of class "scm"
}
\description{
These is a function that fits a GAM model via covariance modelling using \link[mgcv:gam]{mgcv::gam}

These is a function that fits a GAM model via covariance modelling using \link[mgcv:gam]{mgcv::gam}
}
