% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_formula.R
\name{get_foo}
\alias{get_foo}
\title{Processing the model formula provided by the users}
\usage{
get_foo(foo_user, d)
}
\arguments{
\item{foo_user}{formula provided by the user}

\item{d}{dimension of the outcome}
}
\value{
Three list of formulas:one need to be passed to the gam function and the other two are useful for the summary function
}
\description{
This function takes the list of formulas provided by the user, then they are  processed to be passed to the gam function and to be estimated by gam_scm model
}
