% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_formula2.R
\name{get_foo2}
\alias{get_foo2}
\title{This function take the list of formulas provide by the user, then processing it  to be passed to the gam function to estimate gam_scm model}
\usage{
get_foo2(foo_user2, d)
}
\arguments{
\item{d}{dimension of the outcome}

\item{foo_user}{formula provided by the user}
}
\value{
A list of formulas
}
\description{
These is a function that fits a GAM model via covariance modelling using \link[mgcv:gam]{mgcv::gam}
}
