% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{hessian_logm}
\alias{hessian_logm}
\title{Hessian}
\usage{
hessian_logm(eta, y, res)
}
\arguments{
\item{eta}{Linear predictor (n x (d + dx(d+1)/2) matrix).}

\item{y}{Outcome (n x d matrix).}

\item{res}{Memory initialization}
}
\description{
Hessian
}
