% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{internal}
\alias{internal}
\title{Internal function used by the  family (please do not modify this function)}
\usage{
internal()
}
\value{
a list of building functions
}
\description{
Such function masks some functions used internally
}
