% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precision.R
\name{inverse}
\alias{inverse}
\title{Precision matrix (inverse of the covariance matrix)}
\usage{
inverse(X)
}
\arguments{
\item{X}{A covariance matrix}
}
\value{
The precision matrix
}
\description{
The inverse function allows to obtain the precision matrix
}
