% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Labels_mat.R
\name{l_mat}
\alias{l_mat}
\title{Labels of mcd decomposition}
\usage{
l_mat(d)
}
\arguments{
\item{d}{The dimension of the outcome}
}
\value{
A matrix with the labels of the uncostrained mcd elements
}
\description{
This function return the labels of the unconstrained mcd elements
}
\examples{

}
