% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LD.R
\name{LD}
\alias{LD}
\title{LD decomposition from mcd}
\usage{
LD(x, d)
}
\arguments{
\item{x}{The additive predictor vector}

\item{d}{The dimension of the outcome}
}
\value{
The LD matrix via mcd
}
\description{
...
}
