% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ll_logm}
\alias{ll_logm}
\title{Log-likelihood matrix logarithm reparameterization}
\usage{
ll_logm(eta, y)
}
\arguments{
\item{eta}{Linear predictor (n x (d + dx(d+1)/2) matrix).}

\item{y}{Outcome (n x d matrix).}
}
\description{
Log-likelihood matrix logarithm reparameterization
}
