% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lowertri_inv.R
\name{lowertri_inv}
\alias{lowertri_inv}
\title{Lower triangular matrix inversion}
\usage{
lowertri_inv(X)
}
\arguments{
\item{X}{A lower triangular matrix}
}
\value{
the inverse of a lower triangular matrix
}
\description{
...
}
