% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcd.R
\name{mcd}
\alias{mcd}
\title{Modified Cholesky decomposition}
\usage{
mcd(X)
}
\arguments{
\item{X}{A covariance matrix}
}
\value{
A matrix the unconstrained elements of the modified Cholesky decomposition...
}
\description{
The mcd function allows to obtain the unconstrained elements of the modified Cholesky decomposition
}
