% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvn_MCD_internal.R
\name{mvn_mcd_iu}
\alias{mvn_mcd_iu}
\title{Family definition (internal use)}
\usage{
mvn_mcd_iu(d = 2, l2)
}
\arguments{
\item{d}{dimension of the outcome}

\item{l2}{initialization of the hessian matrix (dimension n x no_eta x (no_eta+1)/2), with no_eta = d+d x (d + 1)/2}
}
\value{
family
}
\description{
...
}
