% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nll_logM.R
\name{nll_logm}
\alias{nll_logm}
\title{Negative log-likelihood logM}
\usage{
nll_logm(eta, y)
}
\arguments{
\item{eta}{The linear predictor, a matrix of dimension n x (d+d x (d+1)/2))}

\item{y}{The outcome, that is a matrix n x d}
}
\value{
The function returns the negative log-likelihood according to the logM parameterization
}
\description{
The function returns the negative log-likelihood according to the logM parameterization
}
