% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nll_mcd.R
\name{nll_mcd}
\alias{nll_mcd}
\title{Negative log-likelihood MCD}
\usage{
nll_mcd(eta, y)
}
\arguments{
\item{eta}{The linear predictor, a matrix of dimension n x (d+d x (d+1)/2))}

\item{y}{The outcome, that is a matrix n x d}
}
\value{
The function returns the negative log-likelihood according to the MCD parameterization
}
\description{
The function returns the negative log-likelihood according to the MCD parameterization
}
