% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{operators}
\alias{operators}
\alias{SI}
\alias{LEAST}
\alias{PROD_D}
\alias{MIN_D}
\alias{LEAST_D}
\title{Operator functions for subpixel confusion matrices}
\usage{
SI(s_nk, r_nl)

LEAST(s_nk, r_nl)

PROD_D(sp_nk, rp_nl, k, l)

MIN_D(sp_nk, rp_nl, k, l)

LEAST_D(sp_nk, rp_nl, k, l)
}
\arguments{
\item{s_nk}{grade of membership of pixel n to class k assigned by the
assessed dataset (predicted value)}

\item{r_nl}{grade of membership of pixel n to class l assigned by the
reference dataset (observed value)}

\item{sp_nk}{The overestimation value (predictions minus observations)}

\item{rp_nl}{The underestimation value (observations minus predictions)}

\item{k}{Index in the resulting confusion matrix,
indicating which class was predicted (confusion matrix row).}

\item{l}{Index in the resulting confusion matrix,
indicated which class it was supposed to be attributed to
(confusion matrix column).}
}
\value{
The resulting statistic as a number.
}
\description{
These are the operators, both the agreement (with no prefix) and disagreement
(with a \code{D_} prefix) that were defined by
Silvan-Cardenas and Wang (2008).
}
\examples{
observed = c(X1=0.4, X2=0.3, X3=0.2, X4=0.1)
predicted = c(X1=0.5, X2=0.4, X3=0.1, X4=0)
# X1 agreement
SI(predicted["X1"], observed["X1"])
LEAST(predicted["X1"], observed["X1"])
# X1 disagrement: could be attributed to either X3 or X4
relu <- function(x) sapply(x, function(z) max(0,z))
PROD_D(relu(predicted-observed), relu(observed-predicted), 1, 3)
MIN_D(relu(predicted-observed), relu(observed-predicted), 1, 3)
LEAST_D(relu(predicted-observed), relu(observed-predicted), 1, 3)
# Full SCM
SCM(predicted, observed, disagreement=PROD_D)$P
SCM(predicted, observed, disagreement=MIN_D)$P
SCM(predicted, observed, disagreement=LEAST_D)$P
}
