% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcscmfiltering.R
\name{qcscm.filtering}
\alias{qcscm.filtering}
\title{extract rows of data frame where peer nomination identifier match required pattern of identifier}
\usage{
qcscm.filtering(data, save, file)
}
\arguments{
\item{data}{data frame containing peer nomination (aka identifier) data only, within peer group identifiers are separated by whitespace. Variable names are "sch", "scm_class_nr", "scm_owngrp", "scm_othgrp1", "scm_othgrp2", "scm_othgrp3", "scm_likemst", & ""scm_likelst" (NOTE: variable names without prefix "BA_", "FU1_", or "FU2_"; can be removed by \code{names(df) <- gsub("BA_", "", names(df))})!}

\item{save}{logical. If \code{TRUE}, returned data frame will be saved as .csv with ; separation}

\item{file}{required if \code{save == TRUE}. \code{file} has to be absolute path including file name and appendix .csv in "" where returned data set should be stored.}
}
\value{
data frames with not matching pattern
}
\description{
this function filters for mismatching identifier pattern (4 digits) and extracts rows with least one pattern-mismatching identifier. Identifiers from same peer group will be listed separately column wise.
}
