% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmprep.R
\name{scm.prep}
\alias{scm.prep}
\title{format peer nomination data for external software}
\usage{
scm.prep(data, save, file)
}
\arguments{
\item{data}{data frame containing peer nomination (aka identifier) data only, within peer group identifiers are separated by whitespace. Variable names are "sch", "scm_class_nr", "scm_owngrp", "scm_othgrp1", "scm_othgrp2", & "scm_othgrp3" (NOTE: variable names without prefix "BA_", "FU1_", or "FU2_"; can be removed by \code{names(df) <- gsub("BA_", "", names(df))})!}

\item{save}{logical. If \code{TRUE}, returned data frame will be saved as .txt with ; separation}

\item{file}{required if \code{save == TRUE}. \code{file} has to be absolute path in "" where returned data set should be stored.}
}
\value{
data frames of each school and stream separately
}
\description{
this function prepares peer nomination data for the format required for external software data analysis
}
