% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sigma.R
\name{Sigma_mat}
\alias{Sigma_mat}
\title{Return the covariance matrix in matrix}
\usage{
Sigma_mat(pred_Sigma)
}
\arguments{
\item{pred_Sigma}{A vector with elements needed to fill the covariance matrix}
}
\value{
The function returns the covariance matrix in matrix form
}
\description{
The function returns the covariance matrix
}
