% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spscalc.R
\name{sps.calc}
\alias{sps.calc}
\title{calculate amount of social preferences}
\usage{
sps.calc(data, save, file)
}
\arguments{
\item{data}{data frame containing social preference (aka identifier) data only, within preference identifiers are separated by whitespace. Variable names are "sch", "scm_class_nr", "scm_likemst", & ""scm_likelst" (NOTE: variable names without prefix "BA_", "FU1_", or "FU2_"; can be removed by \code{names(df) <- gsub("BA_", "", names(df))})!}

\item{save}{logical. If \code{TRUE}, returned data frame will be saved as .txt with ; separation}

\item{file}{required if \code{save == TRUE}. \code{file} has to be absolute path in "" where returned data set should be stored.}
}
\value{
extended input data frame containing variable \code{n_likemst} (amount how often identifier listed as most liked) and \code{n_likelst} (amount how often identifier listed as least liked)
}
\description{
this function calculates how often a person (identifier) is listed as liked most and like least within school/stream
}
