% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/totals.scm.R
\name{totals.scm}
\alias{totals.scm}
\title{Calculate total metrics}
\usage{
totals.scm(scm)
}
\arguments{
\item{scm}{The \code{scm} object to enhance.}
}
\value{
An enhanced \code{scm} object.
}
\description{
Accessory function that enhances an input \code{SCM} object's matrices
\code{P} and \code{U} with row and column totals. This makes the totals
easier to read, but the resulting matrix is no longer compatible with
one that is produced by \code{SCM} without enhancement.
}
\examples{
observed = c(Class_1=0.5, Class_2=0.375, Class_3=0.125)
predicted = c(Class_1=0.625, Class_2=0.25, Class_3=0.125)
SCM(predicted, observed, agreement="*", disagreement=NULL)
totals.scm(SCM(predicted, observed, agreement="*", disagreement=NULL))
# Same as:
SCM(predicted, observed, agreement="*", disagreement=NULL, totals=TRUE)
}
