% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_formula_mcd.R
\name{formula_mcd_int}
\alias{formula_mcd_int}
\title{Initialization formula for the covariance model}
\usage{
formula_mcd_int(d)
}
\arguments{
\item{d}{The dimension of the outcome}
}
\value{
The initializing formula (a list) for the covariance model
}
\description{
This function return the initialization formula for the covariance model,
}
\examples{

}
