% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{internal}
\alias{internal}
\title{A pool of internal functions used by the family}
\usage{
internal()
}
\value{
a list of building functions
}
\description{
Such function masks some functions used internally (please do not modify this function)
}
