% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logm.R
\name{logm}
\alias{logm}
\title{Matrix logarithm}
\usage{
logm(X)
}
\arguments{
\item{X}{A covariance matrix}
}
\value{
The logarithm of the covariance matrix
}
\description{
The logm function allows to obtain the unconstrained elements of the matrix logarithm transform
}
