% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvn_scm.R
\name{mvn_scm}
\alias{mvn_scm}
\title{Multivariate gaussian attive model with covariance modelling}
\usage{
mvn_scm(d = 2, nb = 1, param = NULL)
}
\arguments{
\item{d}{dimension of the outcome}

\item{nb}{number of blocks}

\item{param}{type of parameterisation: MCD or logM}
}
\value{
family
}
\description{
...
}
