% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{operators}
\alias{operators}
\alias{SI}
\alias{LEAST}
\alias{PROD_D}
\alias{MIN_D}
\alias{LEAST_D}
\title{Operator functions for subpixel confusion matrices}
\usage{
SI(s_nk, r_nl)

LEAST(s_nk, r_nl)

PROD_D(sp_nk, rp_nl, k, l)

MIN_D(sp_nk, rp_nl, k, l)

LEAST_D(sp_nk, rp_nl, k, l)
}
\arguments{
\item{s_nk}{The matrix of predicted values}

\item{r_nl}{The matrix of observed values}

\item{sp_nk}{The overestimation matrix (predictions minus observations)}

\item{rp_nl}{The underestimation matrix (observations minus predictions)}

\item{k}{Index in the resulting confusion matrix,
indicating which class was predicted (confusion matrix row).}

\item{l}{Index in the resulting confusion matrix,
indicated which class it was supposed to be attributed to
(confusion matrix column).}
}
\description{
These are the operators, both the agreement (with no prefix) and disagreement
(with a \code{D_} prefix) that were defined by
Silvan-Cardenas and Wang (2008).
}
