% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_ecov.R
\name{plotECPD}
\alias{plotECPD}
\title{Heatmap of the Empirical Covariance transformations}
\usage{
plotECPD(ecov, trans = NULL, lab = NULL, di = FALSE, ...)
}
\arguments{
\item{ecov}{empirical covariance matrix}

\item{trans}{an optional function that allows to obtain a transformation of the empirical covariance matrix}

\item{lab}{an optional vector that allows to specify the label of the variables}

\item{di}{logical. If FALSE (default) the diagonal elements are not showed}

\item{...}{optional arguments to pass to hcl.colors of the fields::image.plot function}
}
\value{
The empirical covariance, precision or decomposition matrix.
}
\description{
plotECPD allows to plot the heatmap of the covariance matrix transformations
}
\examples{
A <- matrix(c(1, 0.5, 0.25, 0.5, 2, 0.1, 0.25, 0.1, 3),3,3)
plotECPD(A)
plotECPD(A, trans = inverse)
plotECPD(A, trans = mcd)
plotECPD(A, trans = logm)
}
