% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.scm4.R
\name{read.scm4}
\alias{read.scm4}
\title{import single scm output file with information on participant's & group's nomination, centrality, school/stream}
\usage{
read.scm4(file, ...)
}
\arguments{
\item{file}{path to folder that holds the single be imported scm output data frames. The data frame is a .txt and hold the results from the peer nomination analysis conducted in external software.}
}
\value{
data frame with information on participant's & group's nomination, centrality, school/stream
}
\description{
this function imports a single scm output file, with information on participant's & group's nomination & centrality plus adds columns for school/stream information
}
