% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{SCM}
\alias{SCM}
\title{SCM: A package for multivariate gaussian additive models with smoothing covariance modelling}
\description{
The SCM package provides some important functions to fit the model and to see the model output
}
\section{Mypackage functions}{

gam_scm
mvn_scm
internal
summary.scm
mcd
logm
precision
plot_EPCD
}

