% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sigma.R
\name{Sigma_mat}
\alias{Sigma_mat}
\title{List of predicted Sigma matrix}
\usage{
Sigma_mat(pred_Sigma)
}
\arguments{
\item{pred_Sigma}{the matrix n x d(d+1)/2 of the estimated covariance matrix elements (from predict function with type="response" and excluding the mean vector)}
}
\value{
a list of n elements, each corresponding to the estimated d x d covariance matrix
}
\description{
Such function returns a list of the predicted covariance matrix
}
